refactor(InputManager.kt): Parser와 Validator 객체를 InputManager 내부로 이동
