--- src/main/kotlin/lotto/input/InputValidator.kt
@@ -1,4 +1,0 @@
-package lotto.input
-
-object InputValidator {
-}
